(PC-17816) refactor(profile): move ActionPlanTag from pages to components
